+++ tesler-core/src/main/java/io/tesler/core/util/session/impl/LoginServiceImpl.java
@@ -29,7 +29,7 @@ import io.tesler.core.dto.data.view.ScreenDTO;
 import io.tesler.core.dto.data.view.ScreenResponsibility;
 import io.tesler.core.service.UIService;
 import io.tesler.core.service.ViewService;
-import io.tesler.core.service.impl.ResponsibilitiesService;
+import io.tesler.core.service.ResponsibilitiesService;
 import io.tesler.core.service.impl.UserRoleService;
 import io.tesler.core.util.jackson.CustomObjectMapper;
 import io.tesler.core.util.session.LoginService;